backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
tqdm==4.66.1
alembic==1.12.1
ijson==3.2.3
httpx==0.26.0